News , Payment

+++ Driver_management/App_Code/BLL/Shipment.cs
@@ -23,7 +23,7 @@ namespace BLL
 		public string ShippingStatus { get; set; } // New field
 		public DateTime OrderDate { get; set; }    // New field
 		public int OrderNumber { get; set; }  // New property for order number
-		public int Payment {  get; set; }
+		public decimal Payment {  get; set; }
 
 		public static List<Shipment> GetAll()
 		{
@@ -117,7 +117,8 @@ namespace BLL
 					ShippingStatus = row["ShippingStatus"].ToString(),
 					OrderDate = row["OrderDate"] != DBNull.Value ? Convert.ToDateTime(row["OrderDate"]) : DateTime.MinValue,
 					OrderNumber = row["OrderNumber"] != DBNull.Value ? Convert.ToInt32(row["OrderNumber"]) : 0,
-					Payment =row["Payment"] != DBNull.Value ? Convert.ToInt32(row["Payment"]) : 0,
+					Payment = row["Payment"] != DBNull.Value ? Convert.ToDecimal(row["Payment"]) : 0.0m,
+
 				};
 				shipments.Add(shipment);
 			}
